working on saving solutions
