+++ Solutionizer.Tests/TestData/EmptyVs2019PreviewSolution.sln
@@ -1,13 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
-VisualStudioVersion = 16.0.28803.156
+VisualStudioVersion = 16.0.28809.33
 MinimumVisualStudioVersion = 10.0.40219.1
 Global
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {BB2AB6FA-8028-47EA-99C9-5BFA5DE54864}
+		SolutionGuid = {7B9AAE22-664F-4057-BD77-8C1D45DE135A}
 	EndGlobalSection
 EndGlobal
